added JWT authentication and copleted task 3

+++ backend/middleware/validateProduct.js
@@ -2,7 +2,7 @@ const { body, validationResult } = require('express-validator');
 
 exports.validateProduct = [
     body('name').trim().notEmpty().withMessage('Name is required'),
-    body('price').isFloat({ gt: 0 }).withMessage('Price must be greater than 0'),
+    body('price').toFloat().isFloat({ gt: 0 }).withMessage('Price must be greater than 0'),
 
     (req, res, next) => {
         const errors = validationResult(req);
